Campo album añadido al update de song, corrección del player, eliminación pruebas player de discover y componente player añadido a mainRouter
